Refactored rest.py and some other rest framework stuff into a folder "rest_framework" containing them + seperated public and private views into seperated files
