+++ api/app/signals/apps/questionnaires/views/__init__.py
@@ -1,0 +1,19 @@
+# SPDX-License-Identifier: MPL-2.0
+# Copyright (C) 2021 Gemeente Amsterdam
+from signals.apps.questionnaires.views.private import (
+    PrivateQuestionnaireViewSet,
+    PrivateQuestionViewSet
+)
+from signals.apps.questionnaires.views.public import (
+    PublicQuestionnaireViewSet,
+    PublicQuestionViewSet,
+    PublicSessionViewSet
+)
+
+__all__ = (
+    'PrivateQuestionnaireViewSet',
+    'PrivateQuestionViewSet',
+    'PublicQuestionnaireViewSet',
+    'PublicQuestionViewSet',
+    'PublicSessionViewSet',
+)
